1.All photos are loaded to one photoLocation 2. Separation by folders exists only in DB 3. Photo's identifier is byte[]'s hashcode + dateCreation'shashcode. The identifier is long

+++ src/main/java/afanasievald/repository/PhotoRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 public interface PhotoRepository extends CrudRepository<Photo, Integer> {
     List<Photo> findByFolder(Folder folder);
-    Optional<Photo> findByIdentifier(int identifier);
+    Optional<Photo> findByIdentifier(Long identifier);
 }
